Refactor/server utils (#38)
* refactor: remove empty files, move content to utils.ts

* feat: add index.ts to server utils folder

* refactor: move fn checkExistResource.ts to ErrorHandlerUtil.ts

* refactor: change export default to named export

* refactor: rename imports from "@utils/[...] to @utils"

rename import from @utils to work with new @util path

* chore(tsconfig): rename path @utils/*  to @utils
