+++ server/src/controllers/filesController.ts
@@ -1,17 +1,18 @@
 import { NextFunction, Request, Response } from "express";
 import multer from "multer";
 import fs from "fs";
-import { AppError } from "@utils/ErrorHandlerUtil";
-import { logger } from "@utils/loggerUtil";
-import { alertSoundsPath, musicPath } from "@configs/globalPaths";
-import path from "path";
 import {
+  AppError,
   createDirectory,
   deleteDirectory,
   getListOfDirectoryNames,
-  getListOfMp3InFolder
-} from "@utils/filesManipulateUtil";
+  getListOfMp3InFolder,
+  logger
+} from "@utils";
+import { alertSoundsPath, musicPath } from "@configs/globalPaths";
+import path from "path";
 import { alertSoundPrefix } from "@configs/globalVariables";
+
 const maxFilesAtOnce = 30;
 
 if (!fs.existsSync(musicPath)) {
