+++ server/src/middlewares/checkSearchParamsMiddleware.ts
@@ -1,5 +1,5 @@
 import { RequestQuery } from "@types";
-import { AppError } from "@utils/ErrorHandlerUtil";
+import { AppError } from "@utils";
 import { NextFunction, Request, Response } from "express";
 
 const checkSearchParams = (req: Request<{}, {}, {}, RequestQuery>, res: Response, next: NextFunction) => {
