+++ server/src/services/overlays/overlaysService.ts
@@ -1,8 +1,6 @@
 import { Overlay } from "@models/overlayModel";
 import { OverlayDocument } from "@models/types";
-import { checkExistResource } from "@utils/checkExistResourceUtil";
-import { AppError, handleAppError } from "@utils/ErrorHandlerUtil";
-import { logger } from "@utils/loggerUtil";
+import { AppError, handleAppError, checkExistResource, logger } from "@utils";
 import { FilterQuery, UpdateQuery } from "mongoose";
 import { OverlayCreateData, OverlaysFindOptions, OverlayUpdateData, ManyOverlaysFindOptions } from "./types/";
