+++ server/src/services/streamSessions/streamSessionsService.ts
@@ -3,10 +3,7 @@ import { StreamSessionModel, StreamSessionDocument } from "@models/types";
 import { getMostActiveUsersByMsgs, getMostUsedWord, getMessagesCountByDate } from "@services/messages";
 import { getMostActiveUsersByRedemptions } from "@services/redemptions";
 import { getFollowersCount } from "@services/users";
-import { checkExistResource } from "@utils/checkExistResourceUtil";
-import { handleAppError } from "@utils/ErrorHandlerUtil";
-import { getLastNItemsFromMap } from "@utils/getLastNItemsMapUtil";
-import { logger } from "@utils/loggerUtil";
+import { checkExistResource, handleAppError, getLastNItemsFromMap, logger } from "@utils";
 import { FilterQuery, UpdateQuery } from "mongoose";
 import {
   StreamSessionStatisticOptions,
