+++ server/src/services/tags/tagsService.ts
@@ -4,9 +4,7 @@ import { getChatCommandsCount } from "@services/chatCommands";
 import { getMessageCategoriesCount } from "@services/messageCategories";
 import { getTimersCount } from "@services/timers";
 import { getTriggersCount } from "@services/triggers";
-import { checkExistResource } from "@utils/checkExistResourceUtil";
-import { AppError, handleAppError } from "@utils/ErrorHandlerUtil";
-import { logger } from "@utils/loggerUtil";
+import { checkExistResource, AppError, handleAppError, logger } from "@utils";
 import { FilterQuery, UpdateQuery } from "mongoose";
 import { ManyTagsFindOptions, TagCreateData, TagUpdateData } from "./types";
