+++ server/src/services/widgets/widgetsService.ts
@@ -1,8 +1,6 @@
 import { Widgets } from "@models/widgetsModel";
 import { WidgetsDocument } from "@models/types";
-import { checkExistResource } from "@utils/checkExistResourceUtil";
-import { AppError, handleAppError } from "@utils/ErrorHandlerUtil";
-import { logger } from "@utils/loggerUtil";
+import { AppError, handleAppError, logger, checkExistResource } from "@utils";
 import { FilterQuery, UpdateQuery } from "mongoose";
 import { WidgetCreateData, WidgetsFindOptions, WidgetUpdateData, ManyWidgetsFindOptions } from "./types/";
