+++ server/src/stream/CommandsHandler.ts
@@ -7,8 +7,7 @@ import {
   getOneChatCommand,
   updateChatCommandById
 } from "@services/chatCommands";
-import { commandLogger } from "@utils/loggerUtil";
-import { randomWithMax } from "@utils/randomNumbersUtil";
+import { commandLogger, randomWithMax } from "@utils";
 import type { ClientToServerEvents, InterServerEvents, ServerToClientEvents, SocketData } from "@socket";
 import { Server } from "socket.io";
 import MusicYTHandler from "./MusicYTHandler";
