+++ server/src/stream/StreamHandler.ts
@@ -13,14 +13,13 @@ import type {
 import { Server, Socket } from "socket.io";
 
 import { getCurrentStreamSession, updateStreamSessionById } from "@services/streamSessions";
-import retryWithCatch from "@utils/retryWithCatchUtil";
 
 import CommandsHandler from "./CommandsHandler";
 import TriggersHandler from "./TriggersHandler";
 import LoyaltyHandler from "./LoyaltyHandler";
 import MessagesHandler from "./MessagesHandler";
 import { getConfigs } from "@services/configs";
-import { headLogger, messageLogger } from "@utils/loggerUtil";
+import { headLogger, messageLogger, retryWithCatch } from "@utils";
 import TimersHandler from "./TimersHandler";
 import { ChatUserstate } from "tmi.js";
 import { createUserIfNotExist } from "@services/users";
