--- server/src/utils/removeDifferenceSetUtil.ts
@@ -1,7 +1,0 @@
-const removeDifferenceFromSet = (originalSet: Set<string>, toBeRemovedSet: Set<string>) => {
-  [...toBeRemovedSet].forEach(function (v) {
-    originalSet.delete(v);
-  });
-};
-
-export default removeDifferenceFromSet;
