--- server/src/utils/stringsOperationsUtil.ts
@@ -1,7 +1,0 @@
-export const isValidUrl = (urlString: string) => {
-  try {
-    return Boolean(new URL(urlString));
-  } catch (e) {
-    return false;
-  }
-};
